// ui/lib/api/schemas/nodeDataPatch.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 */
import type { ImageDataPatch } from './imageDataPatch'
import type { MaskDataPatch } from './maskDataPatch'
import type { TextDataPatch } from './textDataPatch'

export type NodeDataPatch =
  | {
      text: TextDataPatch
    }
  | {
      image: ImageDataPatch
    }
  | {
      mask: MaskDataPatch
    }